+++ XuanYu.Core/Scene/ISceneRenderSnapshotSource.cs
@@ -1,0 +1,8 @@
+namespace XuanYu.Core.Scene;
+
+public interface ISceneRenderSnapshotSource
+{
+    SceneRenderSnapshot RenderSnapshot { get; }
+
+    event Action<SceneRenderSnapshot>? RenderSnapshotChanged;
+}